Perspective publish entities error #227

+++ lingvodoc/utils/statistics.py
@@ -15,7 +15,7 @@ import traceback
 from pyramid.response import Response
 from pyramid.view import view_config
 
-from sqlalchemy import and_, BigInteger, cast, extract, func, tuple_
+from sqlalchemy import and_, BigInteger, cast, extract, Float, func, tuple_
 from sqlalchemy.orm import aliased
 from lingvodoc.schema.gql_holders import ResponseError
 
@@ -107,7 +107,7 @@ def stat_perspective(perspective_id, time_begin, time_end, locale_id=2):
         # select
         #   coalesce((additional_metadata #>> '{merge, original_client_id}') :: bigint,
         #     client_id) as entry_client_id,
-        #   coalesce((additional_metadata #>> '{merge, min_created_at}') :: bigint,
+        #   coalesce((additional_metadata #>> '{merge, min_created_at}') :: float,
         #     extract(epoch from created_at)) as entry_created_at
         #   from lexicalentry where
         #     parent_client_id = perspective_client_id and
@@ -121,7 +121,7 @@ def stat_perspective(perspective_id, time_begin, time_end, locale_id=2):
                 LexicalEntry.client_id).label('entry_client_id'),
 
             func.coalesce(
-                cast(LexicalEntry.additional_metadata[('merge', 'min_created_at')].astext, BigInteger),
+                cast(LexicalEntry.additional_metadata[('merge', 'min_created_at')].astext, Float),
                 extract('epoch', LexicalEntry.created_at)).label('entry_created_at'))
 
             .filter(and_(
@@ -185,7 +185,7 @@ def stat_perspective(perspective_id, time_begin, time_end, locale_id=2):
 
                     func.coalesce(
                         cast(Entity.additional_metadata[
-                            ('merge', 'min_created_at')].astext, BigInteger),
+                            ('merge', 'min_created_at')].astext, Float),
                         extract('epoch', Entity.created_at)).label('entity_created_at'),
 
                     PublishingEntity.published, PublishingEntity.accepted)
@@ -231,7 +231,7 @@ def stat_perspective(perspective_id, time_begin, time_end, locale_id=2):
 
                     func.coalesce(
                         cast(Entity.additional_metadata[
-                            ('merge', 'min_created_at')].astext, BigInteger),
+                            ('merge', 'min_created_at')].astext, Float),
                         extract('epoch', Entity.created_at)).label('entity_created_at'),
 
                     PublishingEntity.published, PublishingEntity.accepted)
@@ -446,7 +446,7 @@ def stat_dictionary(dictionary_id, time_begin, time_end, locale_id=None):
                 LexicalEntry.client_id).label('entry_client_id'),
 
             func.coalesce(
-                cast(LexicalEntry.additional_metadata[('merge', 'min_created_at')].astext, BigInteger),
+                cast(LexicalEntry.additional_metadata[('merge', 'min_created_at')].astext, Float),
                 extract('epoch', LexicalEntry.created_at)).label('entry_created_at'),
 
             DictionaryPerspective.state_translation_gist_client_id.label('state_client_id'),
@@ -594,7 +594,7 @@ def stat_dictionary(dictionary_id, time_begin, time_end, locale_id=None):
 
                     func.coalesce(
                         cast(Entity.additional_metadata[
-                            ('merge', 'min_created_at')].astext, BigInteger),
+                            ('merge', 'min_created_at')].astext, Float),
                         extract('epoch', Entity.created_at)).label('entity_created_at'),
 
                     DictionaryPerspective.state_translation_gist_client_id.label('state_client_id'),
@@ -654,7 +654,7 @@ def stat_dictionary(dictionary_id, time_begin, time_end, locale_id=None):
 
                     func.coalesce(
                         cast(Entity.additional_metadata[
-                            ('merge', 'min_created_at')].astext, BigInteger),
+                            ('merge', 'min_created_at')].astext, Float),
                         extract('epoch', Entity.created_at)).label('entity_created_at'),
 
                     DictionaryPerspective.state_translation_gist_client_id.label('state_client_id'),
